agents/risk_scorer/requirements.txt
google-adk==1.18.0
fastapi==0.128.8
uvicorn[standard]==0.39.0
python-dotenv==1.2.1
